+++ dune/typetree/proxynode.hh
@@ -65,7 +65,7 @@ namespace Dune {
        */
       template<std::size_t k,
         typename std::enable_if<lazy_enabled<k>::value, int>::type = 0>
-      auto& child(Dune::index_constant<k> = {})
+      auto& child(index_constant<k> = {})
       {
         return node().proxiedNode().template child<k>();
       }
@@ -75,7 +75,7 @@ namespace Dune {
        * \returns a const reference to the i-th child.
        */
       template<std::size_t k>
-      const auto& child (Dune::index_constant<k> = {}) const
+      const auto& child (index_constant<k> = {}) const
       {
         return node().proxiedNode().template child<k>();
       }
@@ -86,7 +86,7 @@ namespace Dune {
        */
       template<std::size_t k,
         typename std::enable_if<lazy_enabled<k>::value, int>::type = 0>
-      auto childStorage (Dune::index_constant<k> = {})
+      auto childStorage (index_constant<k> = {})
       {
         return node().proxiedNode().template childStorage<k>();
       }
@@ -99,7 +99,7 @@ namespace Dune {
        * \returns a copy of the object storing the i-th child.
        */
       template<std::size_t k>
-      auto childStorage (Dune::index_constant<k> = {}) const
+      auto childStorage (index_constant<k> = {}) const
       {
         return node().proxiedNode().template childStorage<k>();
       }